+++ static/chroot_passwd.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-# change root password to the default of: skybian
-
-printf "skybian\nskybian\n" | passwd root